fix: LC-96 extract before hook to common folder
